changed number validator to properly work with decimal numbers

+++ addon/validators/max.js
@@ -57,7 +57,7 @@ export default Validator.extend({
 	},
 
 	validateNumber: function(value, max) {
-		value = parseInt(value, 10);
+		value = parseFloat(value, 10);
 
 		if(isNaN(value)) {
 			return true;
