+++ addon/validators/min.js
@@ -49,7 +49,7 @@ export default Validator.extend({
 	},
 
 	validateNumber: function(value, min) {
-		value = parseInt(value, 10);
+		value = parseFloat(value, 10);
 
 		if(isNaN(value)) {
 			return true;
